Unfork some hermes-windows changes
