+++ include/hermes/VM/CellKinds.def
@@ -93,7 +93,6 @@ CELL_CLASS(JSBigInt, "BigInt")
 CELL_KIND(BoundFunction)
 CELL_KIND(NativeFunction)
 CELL_KIND(NativeConstructor)
-CELL_KIND(FinalizableNativeConstructor)
 CELL_CLASS(JSCallableProxy, "CallableProxy")
 CELL_KIND(FinalizableNativeFunction)
 CELL_KIND(JSGeneratorFunction)
@@ -117,7 +116,6 @@ CELL_RANGE(ArrayImpl, Arguments, JSArray)
 CELL_RANGE(Callable, BoundFunction, JSFunction)
 CELL_RANGE(CodeBlockFunction, JSGeneratorFunction, JSFunction)
 CELL_RANGE(NativeFunction, NativeFunction, FinalizableNativeFunction)
-CELL_RANGE(NativeConstructor, NativeConstructor, FinalizableNativeConstructor)
 CELL_RANGE(TypedArrayBase, Int8Array, BigUint64Array)
 CELL_RANGE(WeakMapImplBase, JSWeakMap, JSWeakSet)
 CELL_RANGE(
