+++ include/hermes/VM/HostModel.h
@@ -72,74 +72,6 @@ class FinalizableNativeFunction final : public NativeFunction {
   }
 };
 
-/// This class represents a host constructor callable from JavaScript.
-/// This is nearly the same as a NativeConstructor, except an additional
-/// function pointer can do cleanup when the FinalizableNativeFunction is
-/// finalized.
-class FinalizableNativeConstructor final : public NativeConstructor {
-  FinalizeNativeFunctionPtr finalizePtr_;
-
- public:
-  static const CallableVTable vt;
-
-  static constexpr CellKind getCellKind() {
-    return CellKind::FinalizableNativeConstructorKind;
-  }
-  static bool classof(const GCCell *cell) {
-    return cell->getKind() == CellKind::FinalizableNativeConstructorKind;
-  }
-
-  /// Create an instance of FinalizableNativeConstructor.
-  /// \param context the context pointer to be passed to the native function
-  /// \param functionPtr the host function
-  /// \param finalizePtr the finalizer function
-  /// \param prototypeHandle the prototype of constructed object
-  /// \param name the name of the constructor function
-  /// \param paramCount number of parameters (excluding `this`)
-  static CallResult<Handle<FinalizableNativeConstructor>> create(
-      Runtime &runtime,
-      void *context,
-      NativeFunctionPtr functionPtr,
-      FinalizeNativeFunctionPtr finalizePtr,
-      Handle<JSObject> prototypeObjectHandle,
-      SymbolID name,
-      unsigned paramCount);
-
-  void *getContext() {
-    return context_;
-  }
-
-  FinalizableNativeConstructor(
-      Runtime &runtime,
-      Handle<JSObject> parent,
-      Handle<HiddenClass> clazz,
-      void *context,
-      NativeFunctionPtr functionPtr,
-      FinalizeNativeFunctionPtr finalizePtr,
-      CreatorFunction *creator,
-      CellKind targetKind)
-      : NativeConstructor(
-            runtime,
-            parent,
-            clazz,
-            context,
-            functionPtr,
-            creator,
-            targetKind),
-        finalizePtr_(finalizePtr) {}
-
- protected:
-  ~FinalizableNativeConstructor() {
-    finalizePtr_(context_);
-  }
-
-  static void _finalizeImpl(GCCell *cell, GC &) {
-    auto *self = vmcast<FinalizableNativeConstructor>(cell);
-    // Destruct the object.
-    self->~FinalizableNativeConstructor();
-  }
-};
-
 /// When a HostObject is created, this proxy provides the business
 /// logic for the HostObject's implementation.
 class HostObjectProxy : public DecoratedObject::Decoration {
