+++ lib/VM/HostModel.cpp
@@ -77,80 +77,6 @@ CallResult<HermesValue> FinalizableNativeFunction::createWithoutPrototype(
   return selfHandle.getHermesValue();
 }
 
-//===----------------------------------------------------------------------===//
-// class FinalizableNativeConstructor
-
-const CallableVTable FinalizableNativeConstructor::vt{
-    {
-        VTable(
-            CellKind::FinalizableNativeConstructorKind,
-            cellSize<FinalizableNativeConstructor>(),
-            FinalizableNativeConstructor::_finalizeImpl),
-        FinalizableNativeConstructor::_getOwnIndexedRangeImpl,
-        FinalizableNativeConstructor::_haveOwnIndexedImpl,
-        FinalizableNativeConstructor::_getOwnIndexedPropertyFlagsImpl,
-        FinalizableNativeConstructor::_getOwnIndexedImpl,
-        FinalizableNativeConstructor::_setOwnIndexedImpl,
-        FinalizableNativeConstructor::_deleteOwnIndexedImpl,
-        FinalizableNativeConstructor::_checkAllOwnIndexedImpl,
-    },
-    FinalizableNativeConstructor::_newObjectImpl,
-    FinalizableNativeConstructor::_callImpl};
-
-void FinalizableNativeConstructorBuildMeta(
-    const GCCell *cell,
-    Metadata::Builder &mb) {
-  mb.addJSObjectOverlapSlots(
-      JSObject::numOverlapSlots<FinalizableNativeConstructor>());
-  NativeConstructorBuildMeta(cell, mb);
-  mb.setVTable(&FinalizableNativeConstructor::vt);
-}
-
-CallResult<Handle<FinalizableNativeConstructor>>
-FinalizableNativeConstructor::create(
-    Runtime &runtime,
-    void *context,
-    NativeFunctionPtr functionPtr,
-    FinalizeNativeFunctionPtr finalizePtr,
-    Handle<JSObject> prototypeObjectHandle,
-    SymbolID name,
-    unsigned paramCount) {
-  Handle<JSObject> parentHandle =
-      Handle<JSObject>::vmcast(&runtime.functionPrototype);
-
-  FinalizableNativeConstructor *cell =
-      runtime.makeAFixed<FinalizableNativeConstructor, HasFinalizer::Yes>(
-          runtime,
-          parentHandle,
-          runtime.getHiddenClassForPrototype(
-              *parentHandle, numOverlapSlots<FinalizableNativeConstructor>()),
-          context,
-          functionPtr,
-          finalizePtr,
-          creatorFunction<JSObject>,
-          CellKind::JSObjectKind);
-  Handle<FinalizableNativeConstructor> selfHandle =
-      JSObjectInit::initToHandle(runtime, cell);
-
-  ExecutionStatus st = defineNameLengthAndPrototype(
-      selfHandle,
-      runtime,
-      name,
-      paramCount,
-      prototypeObjectHandle,
-      Callable::WritablePrototype::Yes,
-      /*strictMode*/ false);
-  if (st == ExecutionStatus::EXCEPTION) {
-    assert(false && "defineLengthAndPrototype() failed");
-    return ExecutionStatus::EXCEPTION;
-  }
-
-  return selfHandle;
-}
-
-//===----------------------------------------------------------------------===//
-// class HostObject
-
 HostObjectProxy::~HostObjectProxy() {}
 
 const ObjectVTable HostObject::vt{
