+++ npm/package.json
@@ -1,5 +1,5 @@
 {
-  "version": "0.71.0.20220831",
+  "version": "0.71.0.20220906",
   "scripts": {
     "unpack-builds": "node unpack-builds.js",
     "unpack-builds-dev": "node unpack-builds.js --dev",
